+++ src-tauri/src/directory_source/commands.rs
@@ -7,7 +7,7 @@ use tauri::State;
 
 #[tauri::command]
 pub(crate) async fn fetch_file_list_in_directory(
-    directory_id: i64,
+    directory_id: u32,
 ) -> Result<Vec<File>, DwataError> {
     // Find the Directory matching the given folder_id
     match DirectorySource::read_one_by_pk(directory_id, db.deref()).await {
@@ -18,7 +18,7 @@ pub(crate) async fn fetch_file_list_in_directory(
 
 #[tauri::command]
 pub(crate) async fn fetch_file_content_list(
-    directory_id: i64,
+    directory_id: u32,
     relative_file_path: &str,
 ) -> Result<HeterogeneousContentArray, DwataError> {
     // Find the Directory matching the given folder_id
